requirements.txt
gradio==4.1.2
openai==1.1.1
PyYAML==6.0
feedparser==6.0.10
pandas==1.5.3